Fixed group maker endpoing and logic. Code clean up. Removed unsued imports, deleted duplicate code, fixed linting issues.
